+++ selendroid-server/src/main/java/io/selendroid/server/AndroidServlet.java
@@ -15,6 +15,7 @@ package io.selendroid.server;
 
 import io.selendroid.exceptions.StaleElementReferenceException;
 import io.selendroid.server.handler.AddCookie;
+import io.selendroid.server.handler.alert.Alert;
 import io.selendroid.server.handler.CaptureScreenshot;
 import io.selendroid.server.handler.ClearElement;
 import io.selendroid.server.handler.ClickElement;
@@ -67,6 +68,10 @@ import io.selendroid.server.handler.SubmitForm;
 import io.selendroid.server.handler.SwitchWindow;
 import io.selendroid.server.handler.UnknownCommandHandler;
 import io.selendroid.server.handler.Up;
+import io.selendroid.server.handler.alert.AlertAccept;
+import io.selendroid.server.handler.alert.AlertDismiss;
+import io.selendroid.server.handler.alert.AlertSendKeys;
+import io.selendroid.server.model.DefaultSelendroidDriver;
 import io.selendroid.server.model.SelendroidDriver;
 import io.selendroid.util.SelendroidLogger;
 
@@ -81,46 +86,51 @@ public class AndroidServlet extends BaseServlet {
   }
 
   protected void init() {
-    postHandler.put("/wd/hub/session", NewSession.class);
-    getHandler.put("/wd/hub/session/:sessionId", GetCapabilities.class);
+    postHandler.put(  "/wd/hub/session", NewSession.class);
+    getHandler.put(   "/wd/hub/sessions", ListSessions.class);
+    getHandler.put(   "/wd/hub/session/:sessionId", GetCapabilities.class);
     deleteHandler.put("/wd/hub/session/:sessionId", DeleteSession.class);
-    getHandler.put("/wd/hub/session/:sessionId/screenshot", CaptureScreenshot.class);
-    postHandler.put("/wd/hub/session/:sessionId/element", FindElement.class);
-    postHandler.put("/wd/hub/session/:sessionId/element/:id/element", FindChildElement.class);
-    postHandler.put("/wd/hub/session/:sessionId/element/:id/elements", FindChildElements.class);
-    postHandler.put("/wd/hub/session/:sessionId/elements", FindElements.class);
-    postHandler.put("/wd/hub/session/:sessionId/element/:id/click", ClickElement.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/text", GetText.class);
-    getHandler.put("/wd/hub/session/:sessionId/url", GetCurrentUrl.class);
-    postHandler.put("/wd/hub/session/:sessionId/url", OpenUrl.class);
-    postHandler.put("/wd/hub/session/:sessionId/element/:id/value", SendKeys.class);
-    getHandler.put("/wd/hub/session/:sessionId/source", LogElementTree.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/source", LogElement.class);
-    postHandler.put("/wd/hub/session/:sessionId/element/:id/clear", ClearElement.class);
-    getHandler.put("/wd/hub/sessions", ListSessions.class);
-    postHandler.put("/wd/hub/session/:sessionId/timeouts/implicit_wait",
-        SetImplicitWaitTimeout.class);
-    postHandler.put("/wd/hub/session/:sessionId/window", SwitchWindow.class);
-    getHandler.put("/wd/hub/session/:sessionId/window/:windowHandle/size", GetWindowSize.class);
-    postHandler.put("/wd/hub/session/:sessionId/element/:id/submit", SubmitForm.class);
-    postHandler.put("/wd/hub/session/:sessionId/keys", SendKeyToActiveElement.class);
-    getHandler.put("/wd/hub/session/:sessionId/title", GetPageTitle.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/selected", GetElementSelected.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/location", ElementLocation.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/attribute/:name",
-        GetElementAttribute.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/size", GetElementSize.class);
-    postHandler.put("/wd/hub/session/:sessionId/execute", ExecuteScript.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/enabled", GetElementEnabled.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/displayed", GetElementDisplayed.class);
-    getHandler.put("/wd/hub/session/:sessionId/cookie", GetCookies.class);
-    postHandler.put("/wd/hub/session/:sessionId/cookie", AddCookie.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/alert_text", Alert.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/alert_text", AlertSendKeys.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/accept_alert", AlertAccept.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/back", GoBack.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/cookie", GetCookies.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/cookie", AddCookie.class);
     deleteHandler.put("/wd/hub/session/:sessionId/cookie", DeleteCookies.class);
     deleteHandler.put("/wd/hub/session/:sessionId/cookie/:name", DeleteNamedCookie.class);
-
-    postHandler.put("/wd/hub/session/:sessionId/forward", GoForward.class);
-    postHandler.put("/wd/hub/session/:sessionId/back", GoBack.class);
-    postHandler.put("/wd/hub/session/:sessionId/refresh", Refresh.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/dismiss_alert", AlertDismiss.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/element", FindElement.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/elements", FindElements.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/element/:id/attribute/:name", GetElementAttribute.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/element/:id/clear", ClearElement.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/element/:id/click", ClickElement.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/element/:id/displayed", GetElementDisplayed.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/element/:id/element", FindChildElement.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/element/:id/elements", FindChildElements.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/element/:id/enabled", GetElementEnabled.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/element/:id/location", ElementLocation.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/element/:id/name", GetElementTagName.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/element/:id/selected", GetElementSelected.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/element/:id/source", LogElement.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/element/:id/submit", SubmitForm.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/element/:id/text", GetText.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/element/:id/value", SendKeys.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/element/:id/size", GetElementSize.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/execute", ExecuteScript.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/forward", GoForward.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/frame", FrameSwitchHandler.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/keys", SendKeyToActiveElement.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/refresh", Refresh.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/screenshot", CaptureScreenshot.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/source", LogElementTree.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/timeouts/implicit_wait", SetImplicitWaitTimeout.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/title", GetPageTitle.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/url", GetCurrentUrl.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/url", OpenUrl.class);
+    postHandler.put(  "/wd/hub/session/:sessionId/window", SwitchWindow.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/window/:windowHandle/size", GetWindowSize.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/window_handle", GetWindowHandle.class);
+    getHandler.put(   "/wd/hub/session/:sessionId/window_handles", GetWindowHandles.class);
     
     // Advanced Touch API
     postHandler.put("/wd/hub/session/:sessionId/touch/click", SingleTapOnElement.class);
@@ -131,17 +141,17 @@ public class AndroidServlet extends BaseServlet {
     postHandler.put("/wd/hub/session/:sessionId/touch/doubleclick", DoubleTapOnElement.class);
     postHandler.put("/wd/hub/session/:sessionId/touch/longclick", LongPressOnElement.class);
     postHandler.put("/wd/hub/session/:sessionId/touch/flick", Flick.class);
-    getHandler.put("/wd/hub/session/:sessionId/window_handle", GetWindowHandle.class);
-    getHandler.put("/wd/hub/session/:sessionId/window_handles", GetWindowHandles.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/name", GetElementTagName.class);
-    postHandler.put("/wd/hub/session/:sessionId/frame", FrameSwitchHandler.class);
+
+    // Custom extensions to wire protocol
+    getHandler.put( "/wd/hub/-selendroid/:sessionId/screen/brightness", GetScreenState.class);
+    postHandler.put("/wd/hub/-selendroid/:sessionId/screen/brightness", SetScreenState.class);
+    postHandler.put("/wd/hub/session/:sessionId/tap/2", InspectorTap.class);
 
     // currently not yet supported
     getHandler.put("/wd/hub/session/:sessionId/orientation", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/orientation", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/timeouts", UnknownCommandHandler.class);
-    postHandler
-        .put("/wd/hub/session/:sessionId/timeouts/async_script", UnknownCommandHandler.class);
+    postHandler.put("/wd/hub/session/:sessionId/timeouts/async_script", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/execute_async", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/ime/available_engines", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/ime/active_engine", UnknownCommandHandler.class);
@@ -149,24 +159,14 @@ public class AndroidServlet extends BaseServlet {
     postHandler.put("/wd/hub/session/:sessionId/ime/deactivate", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/ime/activate", UnknownCommandHandler.class);
     deleteHandler.put("/wd/hub/session/:sessionId/window", UnknownCommandHandler.class);
-    postHandler.put("/wd/hub/session/:sessionId/window/:windowHandle/size",
-        UnknownCommandHandler.class);
-    postHandler.put("/wd/hub/session/:sessionId/window/:windowHandle/position",
-        UnknownCommandHandler.class);
-    getHandler.put("/wd/hub/session/:sessionId/window/:windowHandle/position",
-        UnknownCommandHandler.class);
-    postHandler.put("/wd/hub/session/:sessionId/window/:windowHandle/maximize",
-        UnknownCommandHandler.class);
+    postHandler.put("/wd/hub/session/:sessionId/window/:windowHandle/size", UnknownCommandHandler.class);
+    postHandler.put("/wd/hub/session/:sessionId/window/:windowHandle/position", UnknownCommandHandler.class);
+    getHandler.put("/wd/hub/session/:sessionId/window/:windowHandle/position", UnknownCommandHandler.class);
+    postHandler.put("/wd/hub/session/:sessionId/window/:windowHandle/maximize", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/element/:id", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/element/active", UnknownCommandHandler.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/equals/:other",
-        UnknownCommandHandler.class);
-    getHandler.put("/wd/hub/session/:sessionId/element/:id/css/:propertyName",
-        UnknownCommandHandler.class);
-    getHandler.put("/wd/hub/session/:sessionId/alert_text", UnknownCommandHandler.class);
-    postHandler.put("/wd/hub/session/:sessionId/alert_text", UnknownCommandHandler.class);
-    postHandler.put("/wd/hub/session/:sessionId/accept_alert", UnknownCommandHandler.class);
-    postHandler.put("/wd/hub/session/:sessionId/dismiss_alert", UnknownCommandHandler.class);
+    getHandler.put("/wd/hub/session/:sessionId/element/:id/equals/:other", UnknownCommandHandler.class);
+    getHandler.put("/wd/hub/session/:sessionId/element/:id/css/:propertyName", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/moveto", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/buttondown", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/buttonup", UnknownCommandHandler.class);
@@ -176,37 +176,25 @@ public class AndroidServlet extends BaseServlet {
     getHandler.put("/wd/hub/session/:sessionId/local_storage", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/local_storage", UnknownCommandHandler.class);
     deleteHandler.put("/wd/hub/session/:sessionId/local_storage", UnknownCommandHandler.class);
-    getHandler
-        .put("/wd/hub/session/:sessionId/local_storage/key/:key", UnknownCommandHandler.class);
-    deleteHandler.put("/wd/hub/session/:sessionId/local_storage/key/:key",
-        UnknownCommandHandler.class);
+    getHandler.put("/wd/hub/session/:sessionId/local_storage/key/:key", UnknownCommandHandler.class);
+    deleteHandler.put("/wd/hub/session/:sessionId/local_storage/key/:key", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/local_storage/size", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/location", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/location", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/local_storage", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/local_storage", UnknownCommandHandler.class);
     deleteHandler.put("/wd/hub/session/:sessionId/local_storage", UnknownCommandHandler.class);
-    getHandler
-        .put("/wd/hub/session/:sessionId/local_storage/key/:key", UnknownCommandHandler.class);
-    deleteHandler.put("/wd/hub/session/:sessionId/local_storage/key/:key",
-        UnknownCommandHandler.class);
+    getHandler.put("/wd/hub/session/:sessionId/local_storage/key/:key", UnknownCommandHandler.class);
+    deleteHandler.put("/wd/hub/session/:sessionId/local_storage/key/:key", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/local_storage/size", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/session_storage", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/session_storage", UnknownCommandHandler.class);
     deleteHandler.put("/wd/hub/session/:sessionId/session_storage", UnknownCommandHandler.class);
-    getHandler.put("/wd/hub/session/:sessionId/session_storage/key/:key",
-        UnknownCommandHandler.class);
-    deleteHandler.put("/wd/hub/session/:sessionId/session_storage/key/:key",
-        UnknownCommandHandler.class);
+    getHandler.put("/wd/hub/session/:sessionId/session_storage/key/:key", UnknownCommandHandler.class);
+    deleteHandler.put("/wd/hub/session/:sessionId/session_storage/key/:key", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/session_storage/size", UnknownCommandHandler.class);
     postHandler.put("/wd/hub/session/:sessionId/log", UnknownCommandHandler.class);
     getHandler.put("/wd/hub/session/:sessionId/log/types", UnknownCommandHandler.class);
-
-    
-    postHandler.put("/wd/hub/session/:sessionId/tap/2", InspectorTap.class);
-    // Custom extensions to wire protocol
-    getHandler.put("/wd/hub/-selendroid/:sessionId/screen/brightness", GetScreenState.class);
-    postHandler.put("/wd/hub/-selendroid/:sessionId/screen/brightness", SetScreenState.class);
   }
 
   private void addHandlerAttributesToRequest(HttpRequest request, String mappedUri) {
@@ -236,6 +224,15 @@ public class AndroidServlet extends BaseServlet {
     Response result = null;
     try {
       addHandlerAttributesToRequest(request, handler.getMappedUri());
+      if (!handler.commandAllowedWithAlertPresentInWebViewMode()) {
+        DefaultSelendroidDriver driver =
+            (DefaultSelendroidDriver) request.data().get(AndroidServlet.DRIVER_KEY);
+        if (driver != null && driver.isAlertPresent()) {
+          result = new SelendroidResponse(handler.getSessionId(), 26, "Unhandled Alert present");
+          handleResponse(request, response, (SelendroidResponse) result);
+          return;
+        }
+      }
       result = handler.handle();
     } catch (StaleElementReferenceException se) {
       try {